refactor: 자동차 검증로직 분리
